// src/types/subscriptions.ts
export enum SubscriptionEnum {
  CONVERSATION_CREATED = 'conversationCreated',
  MESSAGE_CREATED = 'messageCreated',
  MESSAGE_DELETED = 'messageDeleted',
  MESSAGE_UPDATED = 'messageUpdated',
  FRIEND_REMOVED = 'friendRemoved',
  FRIEND_REQUEST_CREATED = 'friendRequestCreated',
  FRIEND_REQUEST_ACCEPTED = 'friendRequestAccepted',
  FRIEND_REQUEST_REJECTED = 'friendRequestRejected',
  FRIEND_REQUEST_CANCEL = 'friendRequestCancel',
}